DB 기술에 독립적인  UserDAO
